refactor: Share more menu beetween public toolbar
